+++ semana10/labeX/labex/src/components/ApplicationForms/ApplicationForms.js
@@ -1,8 +1,18 @@
-import styled from "styled-components";
-import { useTripsLists } from "../../hooks/requests";
+import { useTripsLists, applyToTrip } from "../../hooks/requests";
 import React from "react";
 import { Input, Select } from "@chakra-ui/react";
 import useInput from "../../hooks/useInput";
+import { CountriesList } from "../countrieslist/CountriesList";
+import { useHistory } from "react-router-dom";
+import { goToLastPage } from "../../route/Coordinator";
+import { Button } from "@chakra-ui/react";
+
+import {
+  MainContainer,
+  BoxContainer,
+  ButtonsContainer,
+  Title,
+} from "../../constants/Styled/Styled";
 
 export function ApplicationForm() {
   const [name, setName] = useInput("");
@@ -12,14 +22,15 @@ export function ApplicationForm() {
   const [destination, setDestination] = useInput("");
   const [country, setCountry] = useInput("");
 
+  const history = useHistory();
   const trips = useTripsLists();
 
   return (
-    <div>
+    <MainContainer>
       <Select onChange={setDestination} placeholder="Escolha uma viagem">
         {trips.map((trip) => {
           return (
-            <option key={trip.id} value={trip.name}>
+            <option key={trip.id} value={trip.id}>
               {trip.name}
             </option>
           );
@@ -38,11 +49,31 @@ export function ApplicationForm() {
 
       <Input value={job} onChange={setJob} placeholder="Profissão" />
 
-      <Select
-        value={country}
-        onChange={setCountry}
-        placeholder="Escolha um pais"
+      <CountriesList
+        country={country}
+        setCountry={setCountry}
       />
-    </div>
+
+      <ButtonsContainer>
+        <Button
+          onClick={() => goToLastPage(history)}
+          colorScheme="brand"
+          size="lg"
+          color="#251D44"
+        >
+          Voltar
+          </Button>
+
+        <Button
+          onClick={() => applyToTrip(destination, name, age, text, job, country)}
+          colorScheme="brand"
+          size="lg"
+          color="#251D44"
+        >
+          Enviar
+          </Button>
+
+      </ButtonsContainer>
+    </MainContainer>
   );
 }
