+++ semana10/labeX/labex/src/constants/baseURL.js
@@ -1,0 +1,1 @@
+export const baseURL = "https://us-central1-labenu-apis.cloudfunctions.net/labeX/rodrigo-brezolin-paiva"
