+++ semana10/labeX/labex/src/hooks/requests.js
@@ -1,23 +1,43 @@
-import React, {useState, useEffect} from "react";
+import React, { useState, useEffect } from "react";
 import axios from "axios";
+import {useHistory} from "react-router-dom"
+import { goToAdminHome } from "../route/Coordinator";
+import {baseURL} from '../constants/baseURL';
 
-const baseURL = "https://us-central1-labenu-apis.cloudfunctions.net/labeX/rodrigo-brezolin-paiva"
-const auth = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Im93T2g5ZWo2bW50akZqNUNRMVB4IiwiZW1haWwiOiJhc3Ryb2RldkBnbWFpbC5jb20uYnIiLCJpYXQiOjE1ODk1NjI5MDh9.aB4dNbTCkToXB7pdzEa-tuMa-QbRQDUd93eva4-cec0"
-
-export function useTripsLists(){
+export function useTripsLists() {
     const [trips, setTrips] = useState([])
     const URL = `${baseURL}/trips`
+    const history = useHistory()
 
-    useEffect(() =>{
+    useEffect(() => {
         axios.get(URL)
-        .then((res) => {            
-            setTrips(res.data.trips)
+            .then((res) => {
+                setTrips(res.data.trips)
+            })
+            .catch((err) => {
+                alert(err.response.data.message)
+            })
+    }, [])
+
+    return trips
+}
+
+export const applyToTrip = (destination, name, age, text, job, country) => {
+    const URL = `${baseURL}/trips/${destination}/apply`
+    const body = {
+        "name": name,
+        "age": age,
+        "applicationText": text,
+        "profession": job,
+        "country": country,
+    }
+
+    axios.post(URL, body)
+        .then(() => {
+            alert("Inscrição feita com sucesso")
         })
         .catch((err) => {
             alert(err.message)
         })
-    }, [])
-
-    return trips
-    
 }
+
