+++ semana10/labeX/labex/src/pages/CreateTripPage.js
@@ -7,12 +7,16 @@ import {
     BoxContainer,
     ButtonsContainer,
     Title,
-  } from "./Styled/Styled";
+  } from "../constants/Styled/Styled";
   import { Button, ButtonGroup } from "@chakra-ui/react";
+import useProtectedPage from '../hooks/useProtectedPage';
 
 
  const CreateTripPage = () => {
     const history = useHistory();
+
+    useProtectedPage()
+
     return (
         <MainContainer>
